pyproject.toml
[build-system]
requires = ["setuptools>=82.0.1", "wheel>=0.46.3"]
build-backend = "setuptools.build_meta"

[project]
name = "truepresence-ese"
version = "1.0.0"
description = "TruePresenceESE - Telegram Bot Detection System"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.24.1",
  "questionary>=2.1.1",
  "PyYAML>=6.0.3",
  "fastapi>=0.136.0",
  "uvicorn[standard]>=0.44.0",
  "python-telegram-bot>=22.7",
  "redis>=7.4.0",
  "python-dotenv>=1.2.2",
  "pydantic>=2.13.2",
  "orjson>=3.11.8",
  "aiohttp>=3.13.5",
  "numpy>=2.4.4",
  "tenacity>=9.1.4",
  "psycopg2-binary>=2.9.11",
  "passlib[bcrypt]>=1.7",
  "python-multipart>=0.0.26",
  "httpx>=0.28.1",
  "gunicorn>=25.3.0",
  "requests>=2.33.1",
  "certifi>=2026.2.25",
  "urllib3>=2.6.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=9.0.3",
  "ruff>=0.15.11",
  "mypy>=1.20.1",
]

[project.scripts]
truepresence = "truepresence.main:main"
ese = "ese.cli:app"

[tool.setuptools.packages.find]
include = ["truepresence*", "ese*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"

[tool.ruff.lint]
select = ["B", "E", "F", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
disable_error_code = ["import-untyped"]
warn_unused_configs = true
pretty = true
